serialization impl

+++ Needlefish.Compiler.Tests/NeedlefishFormatter.cs
@@ -2,7 +2,7 @@
 
 internal static partial class NeedlefishFormatter
 {
-    public static void WriteHeader(byte[] buffer, ref int offset, ushort id, bool isOptional = false, bool hasValue = false, bool isArray = false, int arrayLength = 0)
+    public static void WriteHeader(byte[] buffer, ref int offset, ushort id, bool isOptional = false, bool hasValue = false, bool isArray = false, ushort arrayLength = 0)
     {
         Write(buffer, ref offset, id);
